==> analysis/99d_collect_plots.Rmd <==
---
title: "Collect plots"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}

.dir_root <- "analysis"
source(file.path(.dir_root, "00_init.R"))

library(ggpubr)

rs_size <- "full"
preproc <- "yj"
pred_set <- "full"

# Colour-blind friendly palette
cols <- c("#000000", "#E69F00")

```

## Discrimination

```{r load-roc, include = FALSE}

rocs_int <- read_rds(file.path(.dir_plot, rs_size, "rocs_internal_val.rds"))

pred_ext <- read_rds(file.path(.dir_mods, "external", preproc, "all_preds.rds"))
rocs_ext <- map_df(pred_ext, roc_curve, growth, .pred_yes, .id = "model")

lbls <- c("LR", "XGB")
names(lbls) <- str_c(c("lr_", "xgb_"), pred_set)
model_lblr <- labeller(model = lbls)

sel <- . %>% 
  filter(model %in% names(lbls)) %>% 
  mutate(model = factor(model, names(lbls), lbls))

roc_int_lx <- sel(rocs_int)
roc_ext_lx <- sel(rocs_ext)

```

```{r roc_int_ext, echo=FALSE}

g_roc <- 
  ggplot(NULL, aes(1 - specificity, sensitivity)) + 
  geom_abline(intercept = 0, slope = 1, linetype = 2) + 
  scale_x_continuous(labels = scales::percent) + 
  scale_y_continuous(labels = scales::percent) + 
  scale_colour_brewer(palette = "Dark2") + 
  coord_equal(expand = FALSE, xlim = c(0, 1), ylim = c(0, 1)) + 
  labs(
    x = "\n1 - Specificity\n",
    y = "Sensitivity\n"
  ) + 
  theme_bw() + 
  theme(
    panel.spacing = unit(2, "lines"), 
    panel.grid.minor = element_blank()
  )
  
g_roc_lx <- g_roc +
  geom_path(data = roc_int_lx, aes(group = resample), alpha = 0.1) + 
  geom_path(data = roc_ext_lx, aes(colour = model), size = 0.8) + 
  facet_wrap(~ model, ncol = 2) +
  guides(colour = guide_none())
g_roc_lx

```

```{r roc-thresh}

closest_to <- function(roc, tr) {
  roc %>% 
    group_by(model) %>% 
    mutate(diff = .threshold - tr) %>% 
    filter(diff >= 0) %>% 
    arrange(diff) %>% 
    summarise_all(head, 1) %>% 
    mutate(.threshold = tr) %>% 
    select(model, .threshold, specificity, sensitivity)
}

thresh <- bind_rows(
  closest_to(roc_ext_lx, 0.1), 
  closest_to(roc_ext_lx, 0.2), 
  closest_to(roc_ext_lx, 0.3), 
  closest_to(roc_ext_lx, 0.4),
  closest_to(roc_ext_lx, 0.5), 
  closest_to(roc_ext_lx, 0.8)
)

make_point <- function(m, colour) {
  geom_point(
    data = subset(thresh, model == m), 
    size = 3, 
    stroke = if(colour == "white") 1.2 else 0.7, 
    shape = 3, 
    colour = colour,
    show.legend = FALSE
  )
}

roc_pub <- ggplot(
    roc_ext_lx, 
    aes(x = 1 - specificity, y = sensitivity, colour = model)
  ) + 
  geom_path(data = subset(roc_ext_lx, model == "LR")) +
  make_point("LR", "white") + 
  make_point("LR", cols[1]) + 
  geom_path(data = subset(roc_ext_lx, model == "XGB")) +
  make_point("XGB", "white") + 
  make_point("XGB", cols[2]) + 
  geom_text(
    data = subset(thresh, model == "LR"), 
    aes(label = .threshold), 
    colour = "#555555", 
    nudge_x = 0.05,
    nudge_y = -0.05
  ) +
  scale_x_continuous(breaks = 0:5 / 5) + 
  scale_y_continuous(breaks = 0:5 / 5) + 
  scale_colour_manual(values = cols) + 
  coord_equal(expand = FALSE, xlim = c(0, 1), ylim = c(0, 1)) + 
  labs(
    x = "\n1 - Specificity\n",
    y = "Sensitivity\n", 
    colour = "Model:"
  ) + 
  theme_bw() + 
  theme(
    panel.spacing = unit(2, "lines"), 
    panel.grid.minor = element_blank(), 
    legend.position = "top"
  )


clas_pub <- ggplot(roc_ext_lx, aes(.threshold, colour = model)) + 
  geom_line(aes(y = sensitivity, linetype = "TPR")) +
  geom_line(aes(y = 1 - specificity, linetype = "FPR")) + 
  scale_x_continuous(breaks = 0:5 / 5) + 
  scale_y_continuous(breaks = 0:5 / 5) + 
  scale_colour_manual(values = cols) + 
  scale_linetype_manual(values = c(2, 1)) + 
  labs(
    x = "\nRisk threshold\n",
    y = "Proportion\n", 
    colour = "Model:", 
    linetype = "Metric:"
  ) +
  guides(
    colour = guide_none(),
    linetype = guide_legend(reverse = TRUE)
  ) + 
  coord_equal(expand = FALSE, xlim = c(0, 1), ylim = c(0, 1)) + 
  theme_bw() + 
  theme(
    panel.spacing = unit(2, "lines"), 
    panel.grid.minor = element_blank(), 
    legend.position = "top"
  )


ggarrange(roc_pub, clas_pub, ncol = 2, legend = "top")

ggsave(file.path(.dir_plot, "external", preproc, "roc-clas-external.png"),
       width = 7, height = 5, dpi = 300)

```

```{r load-prc, include = FALSE}

prcs_int <- read_rds(file.path(.dir_plot, rs_size, "prcs_internal_val.rds"))

pred_ext <- read_rds(file.path(.dir_mods, "external", preproc, "all_preds.rds"))
prcs_ext <- map_df(pred_ext, pr_curve, growth, .pred_yes, .id = "model")

```

```{r prc, echo=FALSE}

prc_int_lx <- sel(prcs_int)
prc_ext_lx <- sel(prcs_ext)

g_prc <- 
  ggplot(NULL, aes(recall, precision)) + 
  geom_hline(yintercept = 0.365, linetype = 2) + 
  scale_x_continuous(labels = scales::percent) + 
  scale_y_continuous(labels = scales::percent) + 
  scale_colour_viridis_d() + 
  guides(colour = guide_none()) + 
  coord_equal(expand = FALSE, xlim = c(0, 1), ylim = c(0, 1)) + 
  labs(
    x = "\nSensitivity (Recall)\n",
    y = "Positive predictive value (Precision)\n"
  ) + 
  theme_bw() + 
  theme(
    panel.spacing = unit(2, "lines")
  )

g_prc_lx <- g_prc +
  geom_path(data = prc_int_lx, aes(group = resample), alpha = 0.2) + 
  geom_path(data = prc_ext_lx, aes(colour = model), size = 0.8)  + 
  facet_wrap(~ model, ncol = 2, labeller = model_lblr)


```

```{r roc-prc-internal, echo = FALSE}

ggarrange(g_roc_lx, g_prc_lx, ncol = 1, nrow = 2, labels = c("A", "B"))

ggsave(file.path(.dir_plot, "external", preproc, "roc-prc-internal.png"),
       width = 7, height = 7, dpi = 300)

```

```{r external}

# ROCs of full models
roc_ext_full <- rocs_ext %>% filter(model %like% "full")

g_roc_full <- 
  g_roc +
  geom_path(data = roc_ext_full, aes(colour = model), size = 0.8) +
  scale_colour_viridis_d(
    labels = c("LR", "FP", "E-NET", "RF", "XGB"),
    limits = str_c(c("lr", "fp", "net",  "rf", "xgb"), "_full")
  ) + 
  guides(colour = guide_legend("Model:"))

# PRCs of full models
prc_ext_full <- prcs_ext %>% filter(model %like% "full")

g_prc_full <- g_prc +
  geom_path(data = prc_ext_full, aes(colour = model), size = 0.8) +
  scale_colour_viridis_d(
    labels = c("LR", "FP", "E-NET", "RF", "XGB"),
    limits = str_c(c("lr", "fp", "net",  "rf", "xgb"), "_full")
  ) +
  guides(colour = guide_legend("Model:"))


ggarrange(g_roc_full, g_prc_full, ncol = 1, nrow = 2, labels = c("A", "B"), common.legend = TRUE)

```

## Calibration

```{r cal-raw}

pred_ext_lx <- pred_ext %>% 
  bind_rows(.id = "model") %>% 
  sel()

g_cal <- ggplot(NULL, aes(
    x = .pred_yes, 
    y = as.integer(growth == "yes")
  )) + 
  geom_abline(colour = "black", lty = 2) + 
  geom_smooth(method = "loess", colour = "black") + 
  scale_color_viridis_d(begin = 0.5, end = 0.5) +
  scale_alpha_continuous(range = c(0, 0.5)) + 
  facet_wrap(~ model, ncol = 2) + 
  guides(colour = FALSE) + 
  labs(y = "Observed proportion\n") + 
  coord_fixed(xlim = c(0, 1), ylim = c(0, 1), expand = FALSE) + 
  theme_bw() + 
  theme(
    panel.spacing = unit(2, "lines")
  )


g_cal_raw <- (g_cal +
  geom_rug(data = pred_ext_lx %>% filter(growth == "yes"), sides = "tl", alpha = 0.1) + 
  geom_rug(data = pred_ext_lx %>% filter(growth == "no"), sides = "bl", alpha = 0.1) + 
  labs(x = "\nRaw predicted probability\n")
) %+% pred_ext_lx


```

```{r cal-recal}

cal_ext <- read_rds(file.path(.dir_perf, "02_calibration", "external", preproc, "all_cal_preds.rds"))
cal_ext_lx <- cal_ext %>% 
  bind_rows(.id = "model") %>% 
  sel()

g_cal_rec <- (g_cal +
  geom_rug(data = cal_ext_lx %>% filter(growth == "yes"), sides = "tl", alpha = 0.1) + 
  geom_rug(data = cal_ext_lx %>% filter(growth == "no"), sides = "bl", alpha = 0.1) + 
  labs(x = "\nRe-calibrated probability\n")
) %+% cal_ext_lx


```

```{r cal-pub}

ggplot(
    cal_ext_lx, 
    aes(
      x = .pred_yes, 
      y = as.integer(growth == "yes")
    )
  ) + 
  geom_rug(data = cal_ext_lx %>% filter(growth == "yes"), sides = "tl", alpha = 0.1) + 
  geom_rug(data = cal_ext_lx %>% filter(growth == "no"), sides = "bl", alpha = 0.1) + 
  geom_abline(colour = "darkgrey", show.legend = FALSE) + 
  geom_smooth(
    data = pred_ext_lx, 
    aes(colour = "Raw", linetype = "2"), 
    method = "loess", se = FALSE
  ) + 
  geom_smooth(
    aes(colour = "Re-calibrated", linetype = "1"),
    method = "loess",
    fill = "#AAAAAA"
  ) + 
  scale_x_continuous(breaks = 0:5 / 5) + 
  scale_y_continuous(breaks = 0:5 / 5) + 
  scale_color_manual(values = rev(cols)) +
  scale_alpha_continuous(range = c(0, 0.5)) + 
  facet_wrap(~ model, ncol = 2) + 
  labs(
    x = "\nPredicted probability",
    y = "Observed proportion\n", 
    colour = "Probabilites:"
  ) + 
  guides(
    colour = guide_legend(override.aes = list(fill = NA, linetype = c(2, 1))),
    linetype = FALSE
  ) + 
  coord_equal(xlim = c(0, 1), ylim = c(0, 1), expand = FALSE) + 
  theme_bw() + 
  theme(
    panel.spacing = unit(2, "lines"),
    panel.grid.minor = element_blank(), 
    legend.position = "top"
  )

ggsave(file.path(.dir_plot, "external", preproc, "cal-external.png"),
       width = 7, height = 5, dpi = 300)


```


```{r run-all, echo = FALSE}

```
